recodai_luc_scientific_image_forgery_detection/requirements.txt
# Core dependencies
torch>=2.0
torchvision>=0.15
numpy>=1.23
pillow>=9.0

# Machine learning
scikit-learn>=1.3
pytorch-lightning>=2.0.0

# Data processing
pandas>=1.5.0

# Augmentations
albumentations>=1.3.0

# Pre-trained models
timm>=0.9.0
segmentation-models-pytorch>=0.3.0

# Hyperparameter tuning
optuna>=3.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
opencv-python>=4.7.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Kaggle API
kaggle>=1.5.12

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
